add typ controll to the date functions
